requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
langchain-ollama
python-dotenv
sqlalchemy
psycopg2-binary
psycopg
streamlit
requests
faiss-cpu
langchain-community
pypdf
langgraph-checkpoint-postgres
langsmith
unstructured
python-docx
pytesseract
Pillow
